+++ wallets/provider-halo/CHANGELOG.md
@@ -1,3 +1,7 @@
+# [0.27.0](https://github.com/rango-exchange/rango-client/compare/provider-halo@0.26.0...provider-halo@0.27.0) (2024-02-07)
+
+
+
 # [0.26.0](https://github.com/rango-exchange/rango-client/compare/provider-halo@0.25.0...provider-halo@0.26.0) (2024-01-22)